requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.2.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
faiss-cpu==1.10.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
html2text==2024.2.26
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.79
langchain-huggingface==0.1.2
langchain-openai==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.3.45
lxml==5.3.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.8.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
openai==1.64.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
ragas==0.2.13
rank-bm25==0.2.2
RapidFuzz==3.12.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.23.0
